requirements.txt
ginger-dj==5.1
mysqlclient==2.1.1
pyyaml==6.0.1
uritemplate==4.1.1
pre-commit==3.0.1
psycopg2==2.9.5
commitizen==2.42.1
docutils==0.19
coverage==7.2.1
coverage-threshold==0.4.4
redis==4.5.5
prometheus-client==0.20.0
pytz==2024.1
inflection==0.5.1
